feat(remap transform): add `tokenize` function (#4093)
