+++ src/mapping/parser/mod.rs
@@ -508,7 +508,7 @@ mod tests {
     use crate::mapping::query::function::{
         ContainsFn, DowncaseFn, FormatTimestampFn, Md5Fn, NowFn, ParseJsonFn, ParseTimestampFn,
         Sha1Fn, SliceFn, StripWhitespaceFn, ToBooleanFn, ToFloatFn, ToIntegerFn, ToStringFn,
-        ToTimestampFn, TruncateFn, UpcaseFn, UuidV4Fn,
+        ToTimestampFn, TokenizeFn, TruncateFn, UpcaseFn, UuidV4Fn,
     };
 
     #[test]
@@ -1171,6 +1171,13 @@ mod tests {
                 Mapping::new(vec![Box::new(Assignment::new(
                     "foo".to_string(),
                     Box::new(SliceFn::new(Box::new(QueryPath::from("foo")), 0, Some(1))),
+                ))]),
+            ),
+            (
+                ".foo = tokenize(.foo)",
+                Mapping::new(vec![Box::new(Assignment::new(
+                    "foo".to_string(),
+                    Box::new(TokenizeFn::new(Box::new(QueryPath::from("foo")))),
                 ))]),
             ),
         ];
